Переделана разбивка модулей для 5.14
